+++ packages/nextclaw-runtime/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nextclaw/runtime
 
+## 0.4.45
+
+### Patch Changes
+
+- Updated dependencies [5b2a035]
+- Updated dependencies [f09ef09]
+  - @nextclaw/core@0.17.19
+
 ## 0.4.44
 
 ### Patch Changes